Fix ENAMETOOLONG on windows
